UPSTREAM: procfs: fixes pthread cross-thread naming if !PR_DUMPABLE
The PR_DUMPABLE flag causes the pid related paths of the
proc file system to be owned by ROOT. The implementation
of pthread_set/getname_np however needs access to
/proc/<pid>/task/<tid>/comm.
If PR_DUMPABLE is false this implementation is locked out.

This patch installs a special permission function for
the file "comm" that grants read and write access to
all threads of the same group regardless of the ownership
of the inode. For all other threads the function falls back
to the generic inode permission check.

Signed-off-by: Janis Danisevskis <jdanis@google.com>

Change-Id: Iea1acc401017e7a95f3387d2dd324c98873f9165
Git-commit: e950a7f43d8b3638558ea8f92cbfab514cbdf7b6
Git-repo: https://android.googlesource.com/kernel/common
Signed-off-by: Manjunatha Madana <mamanj@codeaurora.org>