ARM64: smp: Fix cpu_up() racing with sys_reboot
ARM64: smp: Fix cpu_up() racing with sys_reboot
 Nothing stops a process from hotplugging in a CPU concurrently
 with a sys_reboot() call. In such a situation we could have
 ipi_cpu_stop() mark a cpu as 'offline' and _cpu_up() ignore the
 fact that the CPU is not really offline and call the
 CPU_UP_PREPARE notifier. When this happens stop_machine code will complain
 that the cpu thread already exists and BUG_ON().

CPU0                      CPU1

 sys_reboot()
 kernel_restart()
 machine_restart()
 machine_shutdown()
 smp_send_stop()
 ...                   ipi_cpu_stop()
		       set_cpu_online(1, false)
                       local_irq_disable()
                       while(1)
        <PREEMPT>
 cpu_up()
 _cpu_up()
 if (!cpu_online(1))
 __cpu_notify(CPU_UP_PREPARE...)

 cpu_stop_cpu_callback()
 BUG_ON(stopper->thread)

This is easily reproducible by hotplugging in and out in a tight
loop while also rebooting.

Since the CPU is not really offline and hasn't gone through the
proper steps to be marked as such, let's mark the CPU as inactive.
This is just as easily testable as online and avoids any possibility
of _cpu_up() trying to bring the CPU back online when it never was
offline to begin with. Based on the similar patchset by for arm
targets 040c163( "ARM: smp: Fix cpu_up() racing with sys_reboot)"

CRs-fixed: 758395
Change-Id: Ia13a3aad8cd6616119a07b5114350591173a5d03
Signed-off-by: Abhimanyu Kapur <abhimany@codeaurora.org>

This reverts commit 5988d0d92e055325ba38e49871c154e9ab03c575.